feat: Add Traditional Chinese localization
There is only one zh language in the current implementation, which is
Simplified Chinese by default. This change adds Traditional Chinese
localization and return it when the locale is zh-Hant. We have to use
the languageCode and scriptCode to differentiate between Simplified and
Traditional Chinese.

The `TraditionalChineseAssetPickerTextDelegate` is not included in the
`assetPickerTextDelegates` list because `AssetPickerTextDelegate` lacks
scriptCode information. The change also adds tests to ensure the correct
delegate is returned for both 'zh' and 'zh-Hant' locales.